+++ packages/api/src/index.ts
@@ -1,3 +1,6 @@
+import { Lexicons } from '@atproto/lexicon'
+import { lexicons as internalLexicons } from './client/lexicons'
+
 export { AtUri } from '@atproto/syntax'
 export {
   BlobRef,
@@ -11,7 +14,6 @@ export * from './types'
 export * from './const'
 export * from './util'
 export * from './client'
-export * from './agent'
 export * from './rich-text/rich-text'
 export * from './rich-text/sanitization'
 export * from './rich-text/unicode'
@@ -20,5 +22,14 @@ export * from './moderation'
 export * from './moderation/types'
 export * from './mocker'
 export { LABELS, DEFAULT_LABEL_SETTINGS } from './moderation/const/labels'
+export { Agent } from './agent'
+
+export { AtpAgent, type AtpAgentOptions } from './atp-agent'
 export { BskyAgent } from './bsky-agent'
-export { AtpAgent as default } from './agent'
+
+/** @deprecated */
+export { AtpAgent as default } from './atp-agent'
+
+// Expose a copy to prevent alteration of the internal Lexicon instance used by
+// the AtpBaseClient class.
+export const lexicons = new Lexicons(internalLexicons)