+++ packages/pds/tests/file-uploads.test.ts
@@ -1,6 +1,6 @@
 import fs from 'fs/promises'
 import { gzipSync } from 'zlib'
-import AtpAgent from '@atproto/api'
+import { AtpAgent } from '@atproto/api'
 import { AppContext } from '../src'
 import DiskBlobStore from '../src/disk-blobstore'
 import * as uint8arrays from 'uint8arrays'